+++ test/Parity Tests/Query Tests/Query Tests.js
@@ -28,6 +28,7 @@ module.exports = function ( Driver )
 		require( './test-suite/Expr Tests.js' )( Driver );
 		require( './test-suite/Comparison Operator Tests.js' )( Driver );
 		require( './test-suite/Path Semantics Tests.js' )( Driver );
+		require( './test-suite/Bitwise and Miscellaneous Query Tests.js' )( Driver );
 		require( './test-suite/Query Rejection Tests.js' )( Driver );
 	} );